pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "redspec"
version = "0.1.0"
description = "Generate Azure architecture diagrams from YAML declarations"
requires-python = ">=3.10"
dependencies = [
    "diagrams>=0.24",
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4", "mypy>=1.10", "httpx>=0.27"]
web = ["fastapi>=0.115", "uvicorn[standard]>=0.30", "jinja2>=3.1", "python-multipart>=0.0.9"]
azure = ["azure-identity>=1.15", "azure-mgmt-resourcegraph>=8.0"]
report = ["reportlab>=4.0"]
all = ["redspec[dev,web]"]

[project.scripts]
redspec = "redspec.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/redspec"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]

[tool.mypy]
packages = ["redspec"]
mypy_path = "src"
